refactor: Anniversary 예외 클래스 exception/code 패키지로 이동
