+++ src/main/java/com/umc/halo/domain/notification/exception/AnniversaryException.java
@@ -1,5 +1,6 @@
 package com.umc.halo.domain.notification.exception;
 
+import com.umc.halo.domain.notification.exception.code.AnniversaryErrorCode;
 import com.umc.halo.global.apiPayload.exception.ProjectException;
 
 public class AnniversaryException extends ProjectException {
